+++ examples/aws/terraform/environments/stg/2.admin/backend.tf
@@ -12,12 +12,12 @@ terraform {
     template = "~> 2.0"
   }
   backend "s3" {
-    profile        = "project-stg"
-    bucket         = "project-iac-state-stg"
-    key            = "admin/terraform.tfstate"
-    region         = "ap-northeast-1"
-    encrypt        = true
-    kms_key_id     = "arn:aws:kms:ap-northeast-1:<account-id>:key/<key-id>"
+    profile = "project-stg"
+    bucket  = "project-iac-state-stg"
+    key     = "admin/terraform.stg.tfstate"
+    region  = "ap-northeast-1"
+    /* encrypt        = true
+    kms_key_id     = "arn:aws:kms:ap-northeast-1:<account-id>:key/<key-id>" */
     dynamodb_table = "project-terraform-state-lock-stg"
   }
 }
@@ -29,11 +29,16 @@ provider "aws" {
 }
 data "aws_caller_identity" "current" {}
 
-# data "terraform_remote_state" "s3" {
-#   backend = "s3"
-#   config = {
-#     bucket = "project-iac-state-stg"
-#     key    = "general/terraform.tfstate"
-#     region = "ap-northeast-1"
-#   }
-# }
+
+###################
+# General State
+###################
+data "terraform_remote_state" "general" {
+  backend = "s3"
+  config = {
+    profile = "${var.project}-${var.env}"
+    bucket  = "${var.project}-iac-state-${var.env}"
+    key     = "general/terraform.${var.env}.tfstate"
+    region  = var.region
+  }
+}